refactor(sa): split aggregate init walk; fix union designator edge cases
Decompose InitializerLowering into designator path, slotted walk, and dual
sinks with shared zero geometry. Delete absorbDesignatedAt so each
designator places one value then resumes. Fix union first-arm flat fill,
nested-brace semantics, nested residual zero, and non-brace designator
resume. Split designated-init functional tests by theme.
